Rename draft interface file
