+++ src/app/draft/models/draft-config.interface.ts
